backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
pillow==11.1.0
psycopg-binary==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
